refactor: merge EditorRoleTest with AdminRoleTest
